move redirect component to app.tsx
